landing and product page frontend work + integration

+++ backend/src/main/java/com/rentalplatform/backend/entity/Review.java
@@ -1,12 +1,18 @@
 package com.rentalplatform.backend.entity;
 
-import jakarta.persistence.*;
+import java.math.BigDecimal;
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.time.OffsetDateTime;
-import java.util.UUID;
 
 @Entity
 @Table(name = "reviews")
@@ -30,7 +36,7 @@ public class Review {
     private UUID revieweeId;
 
     @Column(name = "rating", nullable = false)
-    private Double rating;
+    private BigDecimal rating;
 
     @Column(name = "comment")
     private String comment;
